Move committee formatting
