+++ webservices/rest.py
@@ -47,11 +47,10 @@ from flask.ext import restful
 import flask.ext.restful.representations.json
 from json_encoding import TolerantJSONEncoder
 import logging
-from datetime import datetime
 
 from candidates.resources import CandidateResource, CandidateSearch
+from committees.format import format_committees
 from db import db_conn
-import decoders
 from resources import Searchable, SingleResource
 from totals.resources import TotalResource, TotalSearch
 
@@ -63,196 +62,6 @@ flask.ext.restful.representations.json.settings["cls"] = TolerantJSONEncoder
 
 app = Flask(__name__)
 api = restful.Api(app)
-
-
-# still need to implement year
-def format_committees(self, data, page, fields, year):
-    #return data
-
-    results = []
-    for cmte in data:
-        committee = {}
-        # Keeps track of previous info
-        record = {}
-
-        for api_name, fec_name in self.dimcmte_mapping:
-                if cmte['dimcmte'].has_key(fec_name):
-                    committee[api_name] = cmte['dimcmte'][fec_name]
-
-        # addresses, description,
-        addresses = {}
-        for item in cmte['dimcmteproperties']:
-            # shortcut for formatting
-            if item['expire_date'] is not None:
-                if item['expire_date'] > datetime.now():
-                    expired = False
-                else:
-                    expired = True
-            else:
-                expired = False
-
-            if item.has_key('load_date') and item['load_date'] is not None:
-                committee['load_date'] = item['load_date']
-
-            # address
-            address = {}
-            for api_name, fec_name in self.committee_address_field_mappings:
-                if item.has_key(fec_name) and item[fec_name] is not None and fec_name != 'expire_date':
-                    address[api_name] = item[fec_name]
-
-            if 'expire_date' in fields or '*' in fields or fields == []:
-                address['expire_date'] = item['expire_date']
-
-            if address.has_key('state_full'):
-                address['state_full'] = item['cmte_st_desc'].strip()
-
-            if len(address) > 0:
-                if item['expire_date'] == None:
-                    committee['address'] = address
-                else:
-                    if not record.has_key('address'):
-                        record['address'] = {}
-                    record['address'][item['expire_date']] = address
-
-            # properties table called description in api
-            description = {}
-            for api_name, fec_name in self.properties_field_mapping:
-                if item.has_key(fec_name) and item[fec_name] is not None and fec_name != 'expire_date':
-                    description[api_name] = item[fec_name]
-
-            if item.has_key('cand_pty_affiliation') and item['cand_pty_affiliation'] in decoders.party:
-                description['party_full'] = decoders.party[item['cand_pty_affiliation']]
-
-            if len(description) > 0:
-                if 'expire_date' in fields or '*' in fields:
-                    description['expire_date'] = item['expire_date']
-                if expired == False and len(description) > 0:
-                    committee['description'] = description
-                else:
-                    if not record.has_key('description'):
-                         record['description'] = {}
-                    record['description'][item['expire_date']] = description
-
-            # treasurer
-            treasurer = {}
-            for api_name, fec_name in self.treasurer_field_mapping:
-                if item.has_key(fec_name) and item[fec_name] is not None:
-                    treasurer[api_name] = item[fec_name]
-
-            if len(treasurer) > 0:
-                if 'expire_date' in fields or '*' in fields:
-                    treasurer['expire_date'] = item['expire_date']
-                if expired == False:
-                    committee['treasurer'] = treasurer
-                else:
-                    if not record.has_key('treasurer'):
-                        record['treasurer'] = {}
-                    record['treasurer'][item['expire_date']] = treasurer
-
-            # custodian
-            custodian = {}
-            for api_name, fec_name in self.custodian_field_mapping:
-                if item.has_key(fec_name) and item[fec_name] is not None:
-                    custodian[api_name] = item[fec_name]
-
-            if len(custodian) > 0:
-                if 'expire_date' in fields or '*' in fields:
-                    custodian['expire_date'] = item['expire_date']
-                if expired == False:
-                    committee['custodian'] = custodian
-                else:
-                    if not record.has_key('custodian'):
-                        record['custodian'] = {}
-                    record['custodian'][item['expire_date']] = custodian
-
-            #designation
-            for designation in cmte['dimcmtetpdsgn']:
-                status = {}
-                for api_name, fec_name in self.designation_mapping:
-                    if designation.has_key(fec_name) and fec_name != 'expire_date':
-                        status[api_name] = designation[fec_name]
-
-                if 'expire_date' in fields or '*' in fields or fields == []:
-                    status['expire_date'] = designation['expire_date']
-
-                if designation.has_key('cmte_dsgn') and  decoders.designation.has_key(designation['cmte_dsgn']):
-                    status['designation_full'] = decoders.designation[designation['cmte_dsgn']]
-
-                if designation.has_key('cmte_tp') and decoders.cmte.has_key(designation['cmte_tp']):
-                    status['type_full'] = decoders.cmte[designation['cmte_tp']]
-
-                if len(status) > 0:
-                    if designation['expire_date'] == None:
-                        committee['status'] = status
-                    else:
-                        if not record.has_key('status'):
-                            record['status'] = {}
-                        record['status'][designation['expire_date']] = status
-
-            # candidates associated with committees
-            candidate_dict = {}
-            for cand in cmte['dimlinkages']:
-                candidate ={}
-                if cmte['dimlinkages'] != []:
-                    if cmte['dimlinkages'][0].has_key('dimcand'):
-                        if cmte['dimlinkages'][0]['dimcand'][0]['dimcandproperties'] != []:
-                            # this assumes the most recent is last, which seems to be the case
-                            candidate['candidate_name'] = cmte['dimlinkages'][0]['dimcand'][0]['dimcandproperties'][-1]['cand_nm']
-                        if cmte['dimlinkages'][0]['dimcand'][0]['dimcandoffice'] != []:
-                            # not worried about this changing because they would get a new id if it did
-                            candidate['office_sought'] = cmte['dimlinkages'][0]['dimcand'][0]['dimcandoffice'][0]['dimoffice'][0]['office_tp']
-                            candidate['office_sought_full'] = cmte['dimlinkages'][0]['dimcand'][0]['dimcandoffice'][0]['dimoffice'][0]['office_tp_desc']
-
-
-                for api_name, fec_name in self.linkages_field_mapping:
-                    if cand.has_key(fec_name) and fec_name != 'expire_date' and fec_name != 'cand_id':
-                        candidate[api_name] = cand[fec_name]
-                if candidate.has_key('election_years'):
-                    candidate['election_years']= [int(candidate['election_years'])]
-                if 'expire_date' in fields or '*' in fields or fields == []:
-                    candidate['expire_date'] = cand['expire_date']
-                if 'candidate_id' in fields or 'fec_id' in fields or '*' in fields or fields == []:
-                    candidate['candidate_id'] = cand['cand_id']
-                if candidate.has_key('type'):
-                    candidate['type_full'] = decoders.cmte[candidate['type']]
-                if candidate.has_key('designation'):
-                    candidate['designation_full'] = decoders.designation[candidate['designation']]
-                # add all expire dates and save to committee
-                if len(candidate) > 0:
-                    if not candidate_dict.has_key(cand['cand_id']):
-                        candidate_dict[cand['cand_id']] = candidate
-                    elif candidate.has_key('election_years'):
-                        candidate_dict[cand['cand_id']]['election_years'].append(candidate['election_years'][0])
-
-        # one entry per candidate
-        for cand_id in sorted(candidate_dict):
-            if not committee.has_key('candidates'):
-                committee['candidates'] = []
-            committee['candidates'].append(candidate_dict[cand_id])
-
-        # if there are no current records, add the most recent record to the top level committee information
-        for record_type in record:
-            if not committee.has_key(record_type):
-                key = sorted(record[record_type], key=record[record_type].get, reverse=True)[0]
-                committee[record_type] = record[record_type][key]
-                # removing from archive so it is not double posted
-                del record[record_type][key]
-                # adding additional records to archive newest to oldest
-
-            if len(record[record_type]) > 0 and ('archive' in fields or '*' in fields):
-                if not committee.has_key('archive'):
-                    committee['archive'] = {}
-                for key in sorted(record[record_type], key=record[record_type].get, reverse=True):
-                    if not committee['archive'].has_key(record_type):
-                        committee['archive'][record_type] = []
-                    committee['archive'][record_type].append(record[record_type][key])
-        #name short cut
-        if committee.has_key('description') and committee['description'].has_key('name'):
-            committee['name'] = committee['description']['name']
-        results.append(committee)
-
-    return {'api_version':"0.2", 'pagination':page, 'results': results}
-    #return data
 
 
 class NameSearch(Searchable):
